delete apis are now functional
so the apis are functional, i made it so that when you delete an attendee they are removed from all events. Now i need to make the method for removing a an attendee indivdually from an event, i figured out the api for it tonight but ima leave it for tmr.